api/requirements.txt
fastapi==0.101.0
uvicorn==0.23.2
psycopg2-binary==2.9.7
pydantic==2.1.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.25.2